[CCUBE-1432][SR] Amended the translator out into the src component, amended the code structure according to the changes discussed in the code review, amended the use of reducer for the generate schema, removed the preventDefault, amended the imports, amended the props for styling.
