--- src/util/translator/index.ts
@@ -1,2 +1,0 @@
-export * from "./translator";
-export * from "./helper";
