+++ stories/doc-elements.tsx
@@ -1,12 +1,9 @@
-import { useRef } from "react";
-import { FormBuilder, IFormBuilderMethods } from "src";
+import { FormBuilder } from "src";
 
 export const DocElement = () => {
-    const formBuilderRef = useRef<IFormBuilderMethods>(null);
-
     return (
         <div style={{ height: "100vh", width: "100vw" }}>
-            <FormBuilder ref={formBuilderRef} />
+            <FormBuilder />
         </div>
     );
 };
