+++ stories/translator/doc-elements.styles.tsx
@@ -1,8 +1,8 @@
-import { Color } from "@lifesg/react-design-system";
+import { Color } from "@lifesg/react-design-system/color";
 import styled from "styled-components";
 
 interface IContentWrapperProps {
-    visible?: boolean;
+    $visible?: boolean;
 }
 
 export interface IModeButtonProps {
@@ -12,7 +12,7 @@ export interface IModeButtonProps {
 export const ContentWrapper = styled.div<IContentWrapperProps>`
     height: calc(100vh - 5.1rem);
     width: 100vw;
-    display: ${({ visible = true }) => (visible ? "block" : "none")};
+    display: ${({ $visible = true }) => ($visible ? "block" : "none")};
 `;
 
 export const IconWrapper = styled.div`
